Remove pages RPC subdomains and retired frontend pages
