+++ rpc/public/__init__.py
@@ -2,7 +2,6 @@ from .links.handler import handle_links_request
 from .vars.handler import handle_vars_request
 from .users.handler import handle_users_request
 from .gallery.handler import handle_gallery_request
-from .pages.handler import handle_pages_request
 
 
 HANDLERS: dict[str, callable] = {
@@ -10,5 +9,4 @@ HANDLERS: dict[str, callable] = {
   "vars": handle_vars_request,
   "users": handle_users_request,
   "gallery": handle_gallery_request,
-  "pages": handle_pages_request,
 }
