+++ rpc/service/__init__.py
@@ -6,13 +6,11 @@ Requires ROLE_SERVICE_ADMIN.
 from .roles.handler import handle_roles_request
 from .routes.handler import handle_routes_request
 from .reflection.handler import handle_reflection_request
-from .pages.handler import handle_pages_request
 from .rpcdispatch.handler import handle_rpcdispatch_request
 
 HANDLERS: dict[str, callable] = {
   "roles": handle_roles_request,
   "routes": handle_routes_request,
   "reflection": handle_reflection_request,
-  "pages": handle_pages_request,
   "rpcdispatch": handle_rpcdispatch_request,
 }
